requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sentence-transformers==2.7.0
chromadb==0.4.24
python-dotenv==1.0.1
numpy==1.26.4
httpx
python-dotenv
redis

